requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
langchain
langchain-groq
langchain-core
pdfplumber
python-multipart
streamlit